+++ CHANGELOG.md
@@ -36,6 +36,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Create site overlay directory. #1690
 - Set execute permissions for intermediate directories during `wwctl overlay import --parents`. #1655
 - Fix log output formatting during overlay build.
+- Prevent merging of zero-value net.IP fields. #1710
 
 ## v4.6.0rc1, 2025-01-29